pyproject.toml
[project]
name = "eeic-exp-fpga-adaptive-thresholding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
]
